Move `fn_stream` to its own module
This commit addresses https://github.com/awslabs/smithy-rs/pull/2978#discussion_r1337313976
